requirements.txt
amqp==5.2.0
annotated-types==0.6.0
anyio==3.7.1
arrow==1.3.0
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
binaryornot==0.4.4
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cookiecutter==2.5.0
cryptography==41.0.7
daphne==4.0.0
defusedxml==0.7.1
distro==1.8.0
dj-database-url==2.1.0
Django==4.2.7
django-allauth==0.58.2
django-browser-reload==1.12.1
django-environ==0.9.0
django-pwa==1.1.0
django-tailwind==3.6.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
gunicorn==21.2.0
h11==0.14.0
http_ece==1.2.0
httpcore==1.0.2
httpx==0.25.2
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
Jinja2==3.1.2
kombu==5.3.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
msgpack==1.0.7
oauthlib==3.2.2
openai==1.3.7
packaging==23.2
Pillow==10.1.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
py-vapid==1.9.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.17.2
PyJWT==2.8.0
pyOpenSSL==23.3.0
pypng==0.20220715.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2023.3.post1
pywebpush==1.14.1
PyYAML==6.0.1
qrcode==7.4.2
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
text-unidecode==1.3
tqdm==4.66.1
Twisted==23.10.0
txaio==23.1.1
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.4
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.3.0
zope.interface==6.1
